lazy load instrumentation
