+++ packages/datadog-instrumentations/index.js
@@ -1,54 +1,3 @@
 'use strict'
 
-require('./src/amqplib')
-require('./src/amqp10')
-require('./src/aws-sdk')
-require('./src/bluebird')
-require('./src/bunyan')
-require('./src/cassandra-driver')
-require('./src/connect')
-require('./src/couchbase')
-require('./src/cucumber')
-require('./src/dns')
-require('./src/elasticsearch')
-require('./src/express')
-require('./src/fastify')
-require('./src/find-my-way')
-require('./src/generic-pool')
-require('./src/google-cloud-pubsub')
-require('./src/graphql')
-require('./src/grpc')
-require('./src/hapi')
-require('./src/http')
-require('./src/http2')
-require('./src/ioredis')
-require('./src/jest')
-require('./src/kafkajs')
-require('./src/knex')
-require('./src/koa')
-require('./src/memcached')
-require('./src/microgateway-core')
-require('./src/moleculer')
-require('./src/mongodb-core')
-require('./src/mongoose')
-require('./src/mysql')
-require('./src/mysql2')
-require('./src/mocha')
-require('./src/net')
-require('./src/next')
-require('./src/oracledb')
-require('./src/paperplane')
-require('./src/pino')
-require('./src/pg')
-require('./src/promise')
-require('./src/promise-js')
-require('./src/q')
-require('./src/redis')
-require('./src/restify')
-require('./src/router')
-require('./src/rhea')
-require('./src/sharedb')
-require('./src/tedious')
-require('./src/when')
-require('./src/winston')
-require('./src/limitd-client')
+require('./src/helpers/register')
