+++ packages/datadog-instrumentations/src/helpers/instrumentations.js
@@ -1,0 +1,3 @@
+'use strict'
+
+module.exports = {}
